api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "google-genai>=1.70.0",
    "keyring>=25.7.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
    "watchfiles>=1.1.1",
]
